The tree compiles without a warning
284 warning lines, 124 unique sites, to nil on MinGW64 gcc 16.2.

Eighty-nine of them were missing-field-initializers, and they did not need
eighty-nine edits. They fell out of twenty-five member declarations in nine
headers: GamePlayerInfo::teamId alone accounted for forty, because adding a
field to a struct makes every aggregate initialiser in the tree warn at once.
The fix is a default member initializer of empty braces, which is exactly what
aggregate initialization already does at those call sites, so nothing changes
there. Where it does change something is default construction, and that change
is in the direction CLAUDE.md already asks for under the rule that a member
the hash reads is initialised by the time the object exists. nanoPoint is what
that rule cost to learn; twenty-three of these twenty-five members are
std::optional, where the brace initialiser and value-initialization are the
same nullopt.

The rest went the way the tree already handles them. Seventeen unused
parameters get their name commented out, which is the form used sixty-four
times already and four times in Scene.h itself, two lines below the one that
warned. Twelve sign-compare sites take either getSize, the signed size helper
this tree has used ninety-one times, or an explicit cast where the signed side
is provably non-negative -- a fixed-size array, a loop counter from zero, or a
Grid whose constructor rejects a negative extent. TdfParser compares a code
point against a char and now goes through unsigned char, because casting a
signed char straight to unsigned is wrong for any byte over 127.

Three dangling-reference warnings in GameScene_commands are false positives:
the referent is owned by GameMediaDatabase and outlives the reference. They
are silenced by taking the value instead, which costs two atomic refcount
bumps, the piece's polygons being behind a shared_ptr.

GameSimulation::doProjectileImpact keeps its ImpactType parameter, commented
out rather than deleted. It is genuinely unread, but twenty-one call sites
name it and one of them passes Water where the others pass Normal, so whether
the simulation should be ignoring that distinction is a conformance question
and not a warnings one. docs/MAINTAINABILITY-PLAN.md records it.

+++ src/rwe/LoadingNetworkService.cpp
@@ -160,7 +160,7 @@ namespace rwe
         }
 
         auto messageSize = outerMessage.ByteSizeLong();
-        if (messageSize > getSize(sendBuffer) - 4)
+        if (messageSize > static_cast<unsigned long long>(getSize(sendBuffer) - 4))
         {
             throw std::runtime_error("Message to be sent was bigger than buffer size");
         }
